decompose the datapanel widget to its own file
