+++ kb-data-panel/src/datalist.tsx
@@ -8,8 +8,8 @@ import { WorkspaceObjectInfo } from './workspaceHelper';
 import { DataCard } from './dataCard';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { INotebookTracker, NotebookActions } from '@jupyterlab/notebook';
-import { CodeCell } from '@jupyterlab/cells';
+import { INotebookTracker, NotebookActions, /*Notebook*/ } from '@jupyterlab/notebook';
+import { CodeCell, /*CodeCellModel*/ } from '@jupyterlab/cells';
 
 
 export class DataList extends Widget {
@@ -74,19 +74,32 @@ export class DataList extends Widget {
         console.log(wsId + '/' + objId);
         // const context = this.nbTracker.currentWidget.context;
         const current = this.nbTracker.currentWidget;
-        // current.
-        const { context, content } = current;
+        const { context, content, /*contentFactory*/ } = current;
 
         NotebookActions.insertBelow(content);
         const curCell = this.nbTracker.activeCell as CodeCell;
         curCell.model.value.text = wsId + '/' + objId;
+        curCell.model.metadata.set('kbase', {'foo': 'bar'});
         CodeCell.execute(curCell, context.session);
-        // const model = context.model;
-        // const cell = model.contentFactory.createCodeCell({});
-        // cell.value.text = wsId + '/' + objId;
 
+        /* Order of ops
+         * 1. Set up metadata.
+         * 2. Create cell with metadata (content factory will make a "viewer" cell)
+         *   a. Auto-create code
+         *   b. Other metadata as necessary
+         * 3. Execute it
+         *
+         */
+        // const cellModel = new CodeCellModel({});
+        // cellModel.metadata.set('kbase', {'foo': 'bar'});
+        // const cellOptions = { cellModel, contentFactory };
+
+        // const model = context.model;
+        // const cell = model.contentFactory.createCodeCell(cellOptions);
+
+        // const curCell = this.nbTracker.activeCell;
         // model.cells.insert(0, cell);
-        // cell.
+        // cell.value.text = wsId + '/' + objId;
         // const curCell = this.nbTracker.activeCell as CodeCell;
         // CodeCell.execute(curCell, context.session);
