+++ kbase-extension/lab/workspaces/narrativelab-7590.jupyterlab-workspace
@@ -1,1 +1,1 @@
-{"data":{"notebook:ws.9329.obj.1":{"data":{"path":"ws.9329.obj.1","factory":"Notebook"}},"notebook:ws.10292.obj.1":{"data":{"path":"ws.10292.obj.1","factory":"Notebook"}},"layout-restorer:data":{"main":{"dock":{"type":"tab-area","currentIndex":0,"widgets":["notebook:ws.10292.obj.1","notebook:ws.9329.obj.1"]},"mode":"multiple-document","current":"notebook:ws.10292.obj.1"},"left":{"collapsed":false,"widgets":["filebrowser","running-sessions","command-palette","tab-manager"]},"right":{"collapsed":true,"widgets":[]}},"@jupyterlab/settingeditor-extension:plugin":{"sizes":[0.25,0.75],"container":{"editor":"raw","plugin":"@jupyterlab/notebook-extension:tracker","sizes":[0.5,0.5]}}},"metadata":{"id":"/narrative/lab"}}
+{"data":{"layout-restorer:data":{"main":{"dock":{"type":"tab-area","currentIndex":0,"widgets":["notebook:ws.10292.obj.1","notebook:ws.9329.obj.1"]},"mode":"multiple-document","current":"notebook:ws.10292.obj.1"},"left":{"collapsed":false,"widgets":["filebrowser","running-sessions","command-palette","tab-manager"]},"right":{"collapsed":true,"widgets":[]}},"@jupyterlab/settingeditor-extension:plugin":{"sizes":[0.25,0.75],"container":{"editor":"raw","plugin":"@jupyterlab/notebook-extension:tracker","sizes":[0.5,0.5]}},"notebook:ws.10292.obj.1":{"data":{"path":"ws.10292.obj.1","factory":"Notebook"}},"notebook:ws.9329.obj.1":{"data":{"path":"ws.9329.obj.1","factory":"Notebook"}}},"metadata":{"id":"/narrative/lab"}}
